+++ plone/i18n/normalizer/fr.py
@@ -1,7 +1,7 @@
 # -*- coding: UTF-8 -*-
 
 from plone.i18n.normalizer.interfaces import INormalizer
-from zope.interface import implements
+from zope.interface import implementer
 from plone.i18n.normalizer.base import mapUnicode
 
 # French character mapping
@@ -9,6 +9,7 @@ mapping = {
     339: 'oe'
 }
 
+@implementer(INormalizer)
 class Normalizer(object):
     """
     This normalizer can normalize any unicode string and returns a version
@@ -20,7 +21,6 @@ class Normalizer(object):
       >>> verifyClass(INormalizer, Normalizer)
       True
     """
-    implements(INormalizer)
 
     def normalize(self, text, locale=None, max_length=None):
         """